test-requirements.txt
pytest==2.7.0
pytest-django==2.8.0
pytest-cov==1.8.1
coveralls==0.5
